requirements.txt
altair==5.5.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
cachetools==6.1.0
catalogue==2.0.10
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.15
decorator==5.2.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.0
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.13
narwhals==1.48.0
nest-asyncio==1.6.0
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
preshed==3.0.10
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pyzmq==27.0.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
streamlit==1.47.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
